# project/interface/苍穹外卖-前端接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "苍穹外卖-前端接口",
    "description": "苍穹外卖系统前端接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/admin/employee/login": {
      "post": {
        "summary": "员工登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "密码"
                  },
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  }
                },
                "required": [
                  "password",
                  "username"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "description": "主键值"
                        },
                        "name": {
                          "type": "string",
                          "description": "姓名"
                        },
                        "token": {
                          "type": "string",
                          "description": "jwt令牌"
                        },
                        "userName": {
                          "type": "string",
                          "description": "用户名"
                        }
                      }
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/employee/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "string"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/employee": {
      "post": {
        "summary": "新增员工",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "员工id"
                  },
                  "idNumber": {
                    "type": "string",
                    "description": "身份证"
                  },
                  "name": {
                    "type": "string",
                    "description": "姓名"
                  },
                  "phone": {
                    "type": "string",
                    "description": "手机号"
                  },
                  "sex": {
                    "type": "string",
                    "description": "性别"
                  },
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  }
                },
                "required": [
                  "idNumber",
                  "name",
                  "phone",
                  "sex",
                  "username"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "编辑员工信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "idNumber": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "sex": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "idNumber",
                  "name",
                  "phone",
                  "sex",
                  "username"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "string"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/employee/page": {
      "get": {
        "summary": "员工分页查询",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "员工姓名",
            "required": false,
            "example": "张三",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "msg": {
                      "type": "null"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "records": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "username": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "password": {
                                "type": "string"
                              },
                              "phone": {
                                "type": "string"
                              },
                              "sex": {
                                "type": "string"
                              },
                              "idNumber": {
                                "type": "string"
                              },
                              "status": {
                                "type": "number"
                              },
                              "createTime": {
                                "type": "string",
                                "nullable": true
                              },
                              "updateTime": {
                                "type": "string"
                              },
                              "createUser": {
                                "type": "number",
                                "nullable": true
                              },
                              "updateUser": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "id",
                              "username",
                              "name",
                              "password",
                              "phone",
                              "sex",
                              "idNumber",
                              "status",
                              "createTime",
                              "updateTime",
                              "createUser",
                              "updateUser"
                            ]
                          }
                        }
                      },
                      "required": [
                        "total",
                        "records"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/employee/status/{status}": {
      "post": {
        "summary": "启用、禁用员工账号",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "状态，1为启用 0为禁用",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "员工id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "string"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/employee/{id}": {
      "get": {
        "summary": "根据id查询员工",
        "deprecated": false,
        "description": "",
        "tags": [
          "员工相关接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "员工id",
            "required": true,
            "example": 100,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk",
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiJ9.eyJlbXBJZCI6MSwiZXhwIjoxNzYwMzI5MDg2fQ.HPMaNO4N57B13gEwxOoHF_mlwGaUz92upi-yG5WAktk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "createTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createUser": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "id": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "idNumber": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "password": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "sex": {
                          "type": "string"
                        },
                        "status": {
                          "type": "integer",
                          "format": "int32"
                        },
                        "updateTime": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updateUser": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "username": {
                          "type": "string"
                        }
                      }
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}